+++ NEWS.md
@@ -1,6 +1,7 @@
 # medfateland 2.8.3
 * New function unnest_summary() to extract cell summaries as unnested sf.
 * Required medfate 4.9.0
+* Revision of queries to irsic in add_soilgrids()
 
 # medfateland 2.8.2
 * Backwards compatibility of objects initialized with medfate ver. 4.8.3 and 4.8.4